Version 1.0.3 - Attachment stuff added
Added AttachmentNameOptions and AttachmentNameMenu functions to search for emails with a specific Attachment name. 
Added an option for Attachment Name to the workflow of all searches Added an Attachment Name Only search option.
Added an option for a Pre-Built Suspicious Attachment Types Search, and new functions in that workflow that don't allow for delete. This is for info-gathering only.