minor edit to 1.0.12
changed naming convention for Compliance and eDiscovery searches so that the incrementing integer automatically appended to the end (to ensure a unique name if the same name is used more than once) is preceded by a hyphen.  That way, if a name is specified that ends with a number (like an incident tracking number), there is a clear distinction between the name the user specified and the integer added by SADPhishes.